Changed instructions for Repairing Broken Data Links
The instructions had the students copy their data in the Lab1 folder, then save the Lab1.mxd in the Lab1 folder. In trying to break data links the students were instructed to rename the folder to ~Lab1, but since the data are all in the same folder, that doesn't break the data link.

Instructions updated to have students move Lab1.mxd to the parent of Lab1 folder.